requirements.txt
# Discord Security Bot - Python Dependencies
# Generated: 2026-01-05
# Python Version: 3.8+

# Discord Bot Framework
discord.py>=2.3.0

# Async HTTP Client (for VirusTotal, URLhaus, Google Safe Browsing, Hybrid Analysis APIs)
aiohttp>=3.9.0

# YARA Malware Detection
yara-python>=4.3.0

# Cryptography (AES-256 encryption for quarantine)
cryptography>=41.0.0

# Environment Variables (.env file support)
python-dotenv>=1.0.0
